fixed bad prototype

+++ libft.h
@@ -14,7 +14,7 @@ void	*ft_memmove(void *dst, void *src, size_t num);
 void	*ft_memccpy(void *dst, void *src, int c, size_t num);
 void	*ft_memchr(void *ptr, int value, size_t num);
 int	ft_memcmp(const void *s1, const void *s2);
-int	ft_strlen(const char  *s);
+size_t	ft_strlen(const char  *s);
 char	*ft_strdup(const char *s);
 char	*ft_strcpy(char *dst, const char *src);
 char	*ft_strncpy(char *dst, const char *src, size_t n);
